refactor: move NewOAuth2ServerForTesting to separate file
Move the test helper out of oauth.go into oauth_testing.go to keep
the testing import out of production code.
