+++ internal/security/oauth.go
@@ -16,7 +16,6 @@ import (
 	"slices"
 	"strings"
 	"sync"
-	"testing"
 	"time"
 
 	"github.com/gorilla/sessions"
@@ -159,22 +158,6 @@ func (s *OAuth2Server) currentSettings() *conf.Settings {
 
 // For testing purposes
 var testConfigPath string
-
-// NewOAuth2ServerForTesting creates an OAuth2Server with the provided settings
-// for testing. It publishes the settings as the global test snapshot so that
-// currentSettings() returns the expected values, and registers cleanup to
-// restore the global after the test finishes.
-func NewOAuth2ServerForTesting(tb testing.TB, settings *conf.Settings) *OAuth2Server {
-	tb.Helper()
-	conf.SetTestSettings(settings)
-	tb.Cleanup(func() { conf.SetTestSettings(nil) })
-	return &OAuth2Server{
-		Settings:          settings,
-		authCodes:         make(map[string]AuthCode),
-		accessTokens:      make(map[string]AccessToken),
-		throttledMessages: make(map[string]time.Time),
-	}
-}
 
 // Pre-defined errors for token validation
 var (
